core/query: consolidate index.go and indexer.go
`indexer.go` contained the type definition for `query.Indexer` but all
of the methods and logic were defined in `index.go`. It's hard to
remember which is which.
